requirements.txt
grpcio
grpcio-tools
transformers
datasets
numpy
tqdm
torch